refactor: 디렉토리 변경 default form component역할을 하는 header, content, footer 파일을 common 폴더로 위치 변경
