+++ src/views/components/common/content/content.css
@@ -14,6 +14,8 @@
   grid-template-columns: 1fr 320px;
 }
 
+/* article */
+
 .content-article {
   margin-right: 10px;
   height: inherit;
@@ -26,6 +28,8 @@
   border-color: gray;
   height: inherit;
 }
+
+/* aside */
 
 .content-aside {
   padding: 10px 10px 0 0;
