+++ src/views/components/common/footer/footer.js
@@ -1,4 +1,4 @@
-import Component from '../../common/Component.js';
+import Component from '../Component.js';
 
 export default class Footer extends Component {
   template() {
