+++ src/views/main.html
@@ -13,9 +13,9 @@
       rel="stylesheet"
     />
     <link rel="stylesheet" href="public/global.css" />
-    <link rel="stylesheet" href="components/main/header/header.css" />
-    <link rel="stylesheet" href="components/main/content/content.css" />
-    <link rel="stylesheet" href="components/main/footer/footer.css" />
+    <link rel="stylesheet" href="components/common/header/header.css" />
+    <link rel="stylesheet" href="components/common/content/content.css" />
+    <link rel="stylesheet" href="components/common/footer/footer.css" />
     <script type="module" src="components/main/main.js" defer></script>
   </head>
   <body>
